backend/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
python-multipart>=0.0.9,<1
Pillow>=10.0.0,<12
imagehash>=4.3.0,<5
